pyproject.toml
[project]
name = "banque"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dateutils>=0.6.12",
    "fastapi>=0.116.1",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]
